+++ ext/js/language/ar/arabic-text-preprocessors.js
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import {basicTextProcessorOptions} from '../text-preprocessors.js';
+import {basicTextProcessorOptions} from '../text-processors.js';
 
 const optionalDiacritics = [
     '\u0618', // Small Fatha
